Allow enemies to be shot, and tank to be hit.
Shooting enemies doesn't do much more than increment a points
counter (which is not used anywhere) and remove the enemy.

This does limit the number of enemies. Using a low number of 5
for now just for testing, so that we can start to toy with what
happens when all enemies are cleared.

Hitting the player is crude, because the collision bounds are
a circle which only approximate the size of the tank, and in practice
is a bit bigger. It therefore doesn't take into account tank rotations.

Players hitting the tank doesn't cause any kind of damage as of yet,
but rather just makes the enemy get removed.