Force the tank to travel in an arc when turning.
It mostly did this before, except when going from up to down, or left to right.
In those cases, it flipped instead of nicely turning. This ensures it travels
in an arc by normalising the actual travel velocity after interpolating between
current direction and desired direction.

There is still a little weirdness if the tank is heading true north and you
ask it to go south, but it is natural to want to use your left or right arrows
to guide it a specific direction before turning.