Attempting to get tile map to load, requires bumping mono game extended to alpha version.
This is the culmination of several nights stuffing around with dependencies trying to get
everything to work with the latest MonoGame Content Editor (3.8.1), which required bumping
MonoGame.Extended to an alpha version (is it maintained much any more? I don't understand
why they don't cut a new release with what they have now instead of pointing to alpha
versions from the README). This in turn required bumping MonoGame itself to 3.8.1 because
of a breaking change in the Tiled library.

There were other breaking changes in the alpha's of MonoGame.Extended with regards to
keyboard handling, which required calling "Refresh()" each tick on the Mouse and Keyboard
classes from that library.