Tidy up background map tile rendering. Move edge of world calculations and world size calculations to background map.
In theory, the background map could be different sizes for each level. Thus,
leave it to the background map to decide how big the world is and the Game1
can query it when required to setup the collision detection bounds.

In the future, the tile map will also be responsible for creating collision
objects from objects on the map (e.g. buildings) which is supported by the
Tiled map editor. This notion of the BackgroundMap returning a bunch of
collision objects will support that too.