Handle collisions with objects defined in tile map.
Bullets hitting them dissapear (in future will cause an explosion - maybe even update
the contents of the tile map if the object is destructable).

Tank hitting objects in the tile map will push the tank back to the edge, preventing
the tank from moving through them. This bit is quite jankey at times, and I've put
a hack in here to prevent weird behaviour when the collision system behaves in ways
I didn't expect. Will investigate further in the future in the hope of removing this
hack.