Experiment with path finding using QuikGraph library.
As with the collision detection, it may be nice to roll our own version
eventually as a learning exercise. However at this point, the library in
question, though seemingly stagnant, faithfully implements the A* algorithm.
Furthermore, it also deals with the plumbing required for the algorithm,
such as storing graph data structures.

Performance is terrible because this is all just an experiment, but it
will improve as we decide where to actually perform pathfinding and how often.