Allow diagonal movement when pathfinding.
Supports much better paths, but is more expensive because there is
twice as many edges in the graph now.

Added ability to abort AStar algo early when destination node is
found. By default, it keeps going until all nodes are visited.
This speeds it up exceptionally, but it is still slow when pathfinding
to nodes far away on the other side of the map, or when trying to pathfind
to a node which is unreachable. In these cases all nodes are visited which
is expensive.

Also worth noting that the algorithm is not really AStar, because no
heuristic is being used to guide the search. Thus, it is just searching
all nodes right now. This can be improved by, e.g. using a heuristic
function which prefers vertices that are in the same direction as the
target location. This commit leaves a TODO with a rough (but incomplete)
sketch of what this might look like.

Options for improvements in the future include:
 * Use navigation meshes instead of tiles to identify parts of the map
   which can ben visited. In the tilemap there are large swathes of
   space which don't need to be represented by seperate nodes, increasing
   the cost of the algorithm for no benefit.
 * Being more intelligent about when and how to run pathfinding. For
   example, once a path from an enemy to the tank is discovered, only
   recalculate if the tank position diverges from that used for pathfinding
   by a certain amount. Otherwise, keep following the old path.