Experimenting with different pathfinding approaches via navmeshes.
Navmeshes are defined in the tiled map editor via 'Object Layers'. These are then
processed into AdjacencyGraphs in C# for pathfinding.

So far, the approaches aren't perfect, and struggle with getting from nodes in
the navmesh to the position of the tank or enemy which lie outside the navmesh.
The code in this commit tries to add temporary nodes/edges in the graph to
represent these things, but is a bit clunky.