Performance improvements to pathfinding.
Render the edges that the pathfindign algorithm searched, to see if it
is wasteful or not. This uncovered that the search heuristic was off.

Use a stronger heuristic to guide the search. Although there is a very well
documented description of A*-on-grid-with-diagnoal-movement at this blog:
http://theory.stanford.edu/\~amitp/GameProgramming/Heuristics.html\#heuristics-for-grid-maps
I can't seem to make it work. The searched edges are still casting a wide berth.
After testing, euclidian distance works better, and multiplying that by 2 or 3 or 4
returns even better results with minimal edges visited for most cases.

Cache the result of pathfinding and only perform pathfinding again if required.
If the source or dest location change, but they are still somewhere on the path
that we originally discovered, then don't pathfind again (we can reuse the path).

If attempting to pathfind to somewhere that has an obstacle, instead find the
closest vertex in the pathfinding graph which doesn't have an obstacle before
searching. Otherwise, it will always fail to find a path which si kind of uninteresting.