Add red particle's to enemies after they get hit.
Simulates a blast of blood in the direction of the tanks bullet.
In the future, would need to change based on the type of bullet (if
bullet changes are supported at some point). Would also need to
change the type of particle from red (blood) to shrapnel or other
types depending on the enemy type (e.g. drones or robots of some sort).

As with most things, this is sort of a proof of concept, hard coded
as such at this point.