Always centre camera on tank.
Before was doing a smooth scroll to the tank, which was kind of nice.
However, it was really hard to get it to behave predictably and nicely
when we clamp the viewport to the bounds of the world.

Given it is more important to not run the camera over the end of the screen,
remove this feature until a better solution is found (or leave it permanently
out if it isn't actually wanted - it isn't really a deal breaker).